requirements.txt
numpy>=1.26
pandas>=2.0
scipy>=1.11
matplotlib>=3.7
seaborn>=0.12
networkx>=3.1
python-igraph>=0.11
statsmodels>=0.14
spreg>=1.4
libpysal>=4.7
scikit-learn>=1.3
openpyxl>=3.1
